gh-2990 Extract ExecutionOptionsCellBuilder Part 1
